update Readme + fixes
