+++ src/types/index.ts
@@ -1,0 +1,7 @@
+export const CardTypes = [
+  ['другое', 'other'],
+  ['софт-скил', 'soft'],
+  ['кнопка', 'button'],
+  ['дополнительное', 'additional'],
+  ['хард-скил', 'hard'],
+]
