+++ src/view/modal-order-finish-view.ts
@@ -14,12 +14,14 @@ export default class OrderFinishStepView  {
     this.#container = cloneTemplate('#success') as HTMLElement;
     this.#init();
   }
+
   #init() {
     const button = this.#container.querySelector('.order-success__close') as HTMLButtonElement;
     this.#container.querySelector('.order-success__description').textContent = 
         `Списано ${Array.from(this.#items).reduce((sum, item) => sum + item.price, 0)} синапсов`;
     button.onclick = () => this.#event.emit(Event.ORDER_END);
   }
+  
   get template() {
     return this.#container
   }
