+++ src/view/product-card-buсket-view.ts
